requirements.txt
streamlit
python-dotenv
reportlab
openai
Pillow
mysql-connector-python
werkzeug
google-generativeai
